backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
transformers==4.38.2
datasets==2.18.0
peft==0.9.0
trl==0.7.11
accelerate==0.27.2
bitsandbytes==0.42.0
pandas==2.2.1
openpyxl==3.1.2
tensorboard==2.15.2
wandb==0.16.3
huggingface-hub==0.21.4
scipy==1.12.0
scikit-learn==1.4.1.post1
gradio==4.19.2
torch==2.2.1
numpy==1.26.4
tqdm==4.66.2 